:space_invader: create lists in list home / some tweaks and refactor
