+++ components/ListComponents/index.js
@@ -2,13 +2,11 @@ import React from "react";
 import Link from "next/link";
 import moment from "moment";
 
-import Alert from "shared/Alert";
-import ListNameModal from "shared/ListNameModal";
+import ListNameModal from "components/ListComponents/ListNameModal";
 
-import { addCommasToNumber, createUUID, deepCopyObject } from "lib";
-import { getLocalForageLists, setLocalForageItem } from "lib/localForage";
+import { addCommasToNumber } from "lib";
 
-import { MESSAGE_DELAY, MAX_LIST_ITEMS } from "constants/site";
+import { MAX_LIST_ITEMS } from "constants/site";
 
 import utils from "stylesheets/utils.scss";
 import css from "./ListComponents.scss";
@@ -42,7 +40,7 @@ export const Lists = ({ lists }) =>
   </ul>;
 
 export const ListEmpty = () =>
-  <div className={`${css.empty} `}>
+  <div className={`${css.empty}`}>
     <p>This list is empty.</p>
     <p>
       Add up to 50 items from our{" "}
@@ -56,7 +54,7 @@ export const ListEmpty = () =>
   </div>;
 
 export const ListNote = () =>
-  <div className={css.note}>
+  <div className={`${css.note} col-xs-12`}>
     <p>
       <strong>Note:</strong> You won't see lists created in
       another browser here. To view those lists, open the browser
@@ -65,7 +63,7 @@ export const ListNote = () =>
   </div>;
 
 export const ListsEmpty = () =>
-  <div className={`${css.empty} `}>
+  <div className={`${css.empty}  col-xs-12`}>
     <h2 className={css.contentTitle}>You have no lists</h2>
     <p>
       Create lists from our{" "}
@@ -80,17 +78,25 @@ export const ListsEmpty = () =>
   </div>;
 
 export const ListLoading = () =>
-  <div className={`${css.loading} `}>
+  <div className={`${css.loading}`}>
     <h2 className={css.contentTitle}>Loading</h2><p>Please wait…</p>
   </div>;
 
-export const ListsContent = ({ initialized, lists }) =>
+export const ListsContent = ({ initialized, lists, onCreateList }) =>
   <div className={`${utils.container}`}>
     <div className={`row ${css.wrapper}`}>
       {!initialized && <ListLoading />}
-      {lists.length > 0 && <Lists lists={lists} />}
-      {initialized && lists.length === 0 && <ListsEmpty />}
       <ListNote />
+      {initialized && lists.length === 0 && <ListsEmpty />}
+      {initialized &&
+        <ListNameModal
+          className={css.createList}
+          type="create"
+          value=""
+          onChange={onCreateList}
+          name="Create new list"
+        />}
+      {lists.length > 0 && <Lists lists={lists} />}
     </div>
   </div>;
 
@@ -122,149 +128,3 @@ export const ListCheckbox = ({
       </span>
     </label>
   </li>;
-
-export class CheckableLists extends React.Component {
-  state = { showMessage: "", checkedLists: [], lists: [], initialized: false };
-
-  componentDidMount() {
-    this.getLists();
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (this.state.showMessage !== prevState.showMessage)
-      setTimeout(() => this.setState({ showMessage: "" }), MESSAGE_DELAY);
-  }
-
-  getLists = async () => {
-    const { itemId } = this.props;
-    let lists = await getLocalForageLists();
-    lists.sort((a, b) => b.createdAt - a.createdAt);
-    let checkedLists = [];
-    lists.forEach(l => {
-      if (l.selectedHash[itemId]) checkedLists.push(l.uuid);
-    });
-    this.setState({
-      lists: lists,
-      checkedLists: checkedLists,
-      initialized: true
-    });
-  };
-
-  onCheckList = e => {
-    const element = e.target;
-    let id = element.value;
-    let selected = element.checked;
-    if (selected) {
-      this.addItemToList(id);
-    } else {
-      this.removeItemFromList(id);
-    }
-  };
-
-  addItemToList(id) {
-    const { itemId } = this.props;
-    let theList = deepCopyObject(
-      this.state.lists.filter(l => l.uuid === id)[0]
-    );
-    let checkedLists = deepCopyObject(this.state.checkedLists);
-    if (checkedLists.indexOf(id) !== -1 && theList.selectedHash[itemId]) return; // check if item already selected
-    checkedLists.push(id);
-    theList.selectedHash[itemId] = itemId;
-    this.updateList(id, theList, checkedLists, "Item added");
-  }
-
-  removeItemFromList(id) {
-    const { itemId } = this.props;
-    let theList = deepCopyObject(
-      this.state.lists.filter(l => l.uuid === id)[0]
-    );
-    let checkedLists = deepCopyObject(this.state.checkedLists);
-    if (checkedLists.indexOf(id) === -1 && !theList.selectedHash[itemId])
-      return; // check if item not selected
-    checkedLists.splice(checkedLists.indexOf(id), 1);
-    delete theList.selectedHash[itemId];
-    this.updateList(id, theList, checkedLists, "Item removed");
-  }
-
-  updateList = async (uuid, list, checkedLists, message) => {
-    const updatedAt = Date.now();
-    list.updatedAt = updatedAt;
-    list.count = Object.keys(list.selectedHash).length;
-    let lists = deepCopyObject(this.state.lists.filter(l => l.uuid !== uuid));
-    lists.push(list);
-    lists.sort((a, b) => b.createdAt - a.createdAt);
-    this.setState({
-      checkedLists: checkedLists,
-      lists: lists,
-      showMessage: message
-    });
-    await setLocalForageItem(uuid, list);
-  };
-
-  onNameChange = value => {
-    this.createList(value);
-  };
-
-  createList = async listName => {
-    // add the new list to local storage
-    const { itemId } = this.props;
-    const uuid = createUUID();
-    const createdAt = Date.now();
-    let checkedLists = deepCopyObject(this.state.checkedLists);
-    checkedLists.push(uuid);
-    let newLists = deepCopyObject(this.state.lists);
-    let savedList = {
-      name: listName,
-      selectedHash: { [itemId]: itemId },
-      createdAt: createdAt,
-      updatedAt: createdAt
-    };
-    newLists.push(
-      Object.assign({}, savedList, {
-        uuid: uuid,
-        count: 1
-      })
-    );
-    newLists.sort((a, b) => b.createdAt - a.createdAt);
-    await setLocalForageItem(uuid, savedList);
-    this.setState({
-      showMessage: "List created and item added",
-      checkedLists: checkedLists,
-      lists: newLists
-    });
-  };
-
-  render() {
-    const { lists, showMessage, checkedLists } = this.state;
-    const listOfLists = lists.length > 0
-      ? <ul className={css.listOfLists}>
-          {lists.map((l, index) => {
-            const isChecked = checkedLists.indexOf(l.uuid) !== -1;
-            const shouldDisable = l.count >= 50 && !isChecked;
-            return (
-              <ListCheckbox
-                key={`l_${index}`}
-                list={l}
-                shouldDisable={shouldDisable}
-                isChecked={isChecked}
-                onCheckList={this.onCheckList}
-              />
-            );
-          })}
-        </ul>
-      : null;
-    return (
-      <div className="itemLists">
-        <Alert showMessage={showMessage} />
-        <ListNameModal
-          className={css.newListModal}
-          type="create"
-          value=""
-          onChange={this.onNameChange}
-          name="Add to a new list"
-        />
-        {listOfLists}
-      </div>
-    );
-  }
-}
