+++ components/shared/ListView/ListView.scss
@@ -66,6 +66,7 @@
       $grayBorderColor 10px,
       $grayBorderColor 20px
     );
+    opacity: 0.6;
   }
 
   &:last-of-type {
