+++ pages/lists/index.js
@@ -4,7 +4,8 @@ import MainLayout from "components/MainLayout";
 import { ListsContent } from "components/ListComponents";
 import FeatureHeader from "shared/FeatureHeader";
 
-import { getLocalForageLists } from "lib/localForage";
+import { createUUID, deepCopyObject } from "lib";
+import { getLocalForageLists, setLocalForageItem } from "lib/localForage";
 import { LISTS_TITLE } from "constants/lists";
 
 class ListsPage extends React.Component {
@@ -23,6 +24,34 @@ class ListsPage extends React.Component {
     });
   };
 
+  onCreateList = value => {
+    this.createList(value);
+  };
+
+  createList = async listName => {
+    // add the new list to local storage
+    const uuid = createUUID();
+    const createdAt = Date.now();
+    let newLists = deepCopyObject(this.state.lists);
+    newLists.push({
+      uuid: uuid,
+      name: listName,
+      count: 0,
+      createdAt: createdAt
+    });
+    newLists.sort((a, b) => b.createdAt - a.createdAt);
+    let savedList = {
+      name: listName,
+      selectedHash: {},
+      createdAt: createdAt,
+      updatedAt: createdAt
+    };
+    this.setState({
+      lists: newLists
+    });
+    await setLocalForageItem(uuid, savedList);
+  };
+
   render() {
     const { url } = this.props;
     const { lists, initialized } = this.state;
@@ -30,7 +59,11 @@ class ListsPage extends React.Component {
       <MainLayout route={url} pageTitle={LISTS_TITLE}>
         <div id="main" role="main">
           <FeatureHeader title={LISTS_TITLE} />
-          <ListsContent lists={lists} initialized={initialized} />
+          <ListsContent
+            lists={lists}
+            initialized={initialized}
+            onCreateList={this.onCreateList}
+          />
         </div>
       </MainLayout>
     );
